requirements.txt
pygame
RPi.GPIO